Reorganizados componentes
