+++ src/components/actividades/ConfirmarEntrada.js
